Refactor arraytype dispatch into generic Oid dispatch
The arraytypes created during upgrading to 5.0 were previously
emitted on the QD in a bespoke dumpfile for manual dispatching
to the QEs. During the work on support external partitions, it
became necessary to include more Oid synchronization in this
dispatch mechanism. Refactor the code to provide a generic Oid
dispatching mechanism for different types of objects.

Check functions can register dispatch objects in the per-dbinfo
dispatch array, with or without Oids. Objects without Oids will
be resolved as a separate step once the new cluster is restored.