Support external partitions in pg_upgrade
Upgrading a cluster which has external tables as partitions, external
partitions, is now supported whereas it previously made pg_upgrade
error out and fail. The reason why external partitions require special
handling is due to the fact that the partitioning system catalogs are
not distributed to the segment servers. This means that partitioning
is only supported as utility commands on the QD since it requires the
catalogs. Since pg_upgrade operate on a per-segment basis in utility
mode, this presents challenges. Further, external partitions cannot be
defined as part of the table definition for partitioned tables, they
instead need to be swapped into the partition hierarchy via ALTER TABLE
EXCHANGE commands.

This patch implements support for external partitions in pg_upgrade by
building on top of scaffolding added in previous commits:

* The partitioning hierarchy is created intact by adding placeholder
  heap tables in lieu of the external partition, the external table
  is created separataly

* Oid synchronization for the objects are done by registering the new
  table (and associated types) as dispatch objects in pg_upgrade

* ALTER TABLE EXCHANGE commands are stored in a post-processing file
  as they cannot be applied during the upgrade

Additionally, there were pre-existing minor bugs in pg_dump with dumping
external partitons, which are fixed in this patch as well since they
were all intertwined:

* When a partitioning hierarchy containing anonymous partitions, has
  external table(s) exchanged into it, the ALTER TABLE EXCHANGE command
  in the dump must use the rank rather than the partition name (which
  doens't exist for anonymous partitions). Before this patch the dump
  contained this command for the hierarchy which is syntactically wrong:

  ALTER TABLE t EXCHANGE PARTITION "" WITH TABLE tt WITHOUT VALIDATION;

* Fix potential NAMEDATALEN overflow in the partition exchange. Since
  the generated partition tablename is postfixed with a token in the
  dump, the actual name can be hard to predict to a safe level during
  table creation. Ensure that the dump doesn't create broken names that
  could trigger an error during restore.