Remove dead-simple, and therefore redundant, tests.
We have plenty of INSERTs with constants all over the normal test suite.
For example, in the 'insert' test. For the AO and AOCS tables, in
'appendonly' and 'aocs' tests, and the 'uao_dml' tests.

For the partitioned cases, we have inserts in the 'partition' and
'partition1' tests, for example.