Move "CTE functional" tests from TINC to main test suite.
I had to jump through some hoops to run the same test queries with
optimizer_cte_inlining both on and off, like the original tests did. The
actual tests queries are in qp_with_functional.sql, which is included by
two launcher scripts that set the options. The launcher tests are put in
the test schedule, instead of qp_with_functional.sql.

When ORCA is not enabled, the tests are run with gp_cte_sharing on and off.
That's not quite the same as inlining, but it's similar in nature in that
it makes sense to run test queries with it enabled and disabled. There were
some tests for that in with_clause and qp_with_clause tests already, but I
don't think some extra coverage will hurt us.

This is just a straightforward conversion, there might be overlapping tests
between these new tests and existing 'with_clause', 'qp_with_clause', and
upstream 'with' tests. We can clean them up later; these new tests run in
a few seconds on my laptop, so that's not urgent.

A few tests were tagged with "@skip" in the original tests. Test queries
58, 59, and 60 produced different results on different invocations,
apparently depending on the order that a subquery returned rows (OPT-2497).
I left them in TINC, as skipped tests, pending a decision on what to do
about them. Queries 28 and 29 worked fine, so I'm not sure why they were
tagged as "@skip OPT-3035" in the first place. I converted them over and
enabled like the other tests.