Prefer index scans in partition pruning tests.
The original TINC tests where these were copied for tested specifically
that ORCA can use Dynamic Index Scans for these queries. To be closer to
that spirit, even when running without ORCA, try to use index scans when
possible. It exercises more codepaths in the planner anyway, and allows
us to notice more easily if the planner loses the ability to use an index
for some reason.