Parse and check match commands in atmsort
Blocks started by start_matchignore and start_matchsubs could be
closed by either end_ tag, and only the closing command determined
how the content was parsed/used.  This could lead to subtle issues
as only a more-or-less silent warning was issued in the diff.
Instead, save state from the start_ and match the end_ commands to
ensure they match and fail if they don't.