Disconnect OID and relfilenode allocations with new relfilenode counter
The master allocates an OID and provides it to segments during
dispatch. The segments then check if they can use this OID as its
relfilenode. If a segment cannot use the preassigned OID as the
relation's relfilenode, it will generate a new relfilenode value via
nextOid counter. This can result in a race condition between the
generation of the new OID and the segment file being created on disk
after being added to persistent tables. To combat this race condition,
we have a small OID cache... but we have found in testing that it was
not enough to prevent the issue.

To fully solve the issue, we disconnect OID and relfilenode on QE
segments by introducing a nextRelfilenode counter which is similar to
the nextOid counter. QE segments will now use the preassigned OID from
the QD for only relation OID and generate a relfilenode value from the
nextRelfilenode counter. The QD will always use its generated OID as
its relfilenode value.

Reference gpdb-dev thread:
https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/lv6Sb4I6iSI