Remove tests for corrupt pg_aoseg or pg_aocsseg.
It's a good thing that we have a sanity check for this kind of corruption
in the system, but I don't think we necessarily need to have automated
regression tests for a pg_aoseg / gp_relation_node inconsistency. It'd be
more interesting to test that various operations don't lead to an
inconsistency in the first place (and I suppose the rest of the whole test
suite also tests for that implicitly), but we generally don't have explicit
tests for all the various sanity checks we have in the system.