Remove uninteresting tests from qp_misc_rio
This removes uninteresting tests, and tests covering functionality
which is tested elsewhere. The superflous search_path setting is
removed, the amount of tables created reduced and the test data size
reduced. The resource_queue test was more or less duplicated already
but had enough interesting bits to warrant a move to the appropriate
suite.

This shaves roughly 20% of the runtime of tbe test suite.