Properly set attribute inheritance for external tables
External tables doesn't support attribute inheritance, and while
there is no syntax to trick plain inheritance into happening it
was possible to get inherited attributes by injecting an external
table into a partition hierarchy via partition exchange. Commit
fcdd85390 attempted to fix this but was erroneous and subsequently
reverted. The problem with that commit was that it broke off the
inheritance chain too early, the partition still need inheritance
to become part of the hierarchy. This commit essentially does the
same thing but deeper into the inheritance code such that it only
affects the attislocal attribute in pg_attribute.

Also extend the external_tables ICW suite with a test covering this
path.