Add a basic file based progress report to pg_upgrade
In order for higher-level wrappers around pg_upgrade to know what
pg_upgrade does, this adds a trivial file based queue which reports
information on the operations performed. When started with -X, or
--progress, pg_upgrade will write progress as a set of operations
to a logfile. At intervals of either X number of operations, or Y
number of seconds, the file will be released and a new created such
that the released file can be consumed by a wrapper program. The
files will be named <sequence>.inprogress until released when they
are renamed to <sequence>.done. The sequence is gapless. Once the
file is renamed to <sequence>.done, pg_upgrade will never touch it
again and it can at that point be removed.

The format of the progress file is a simple semicolon delimited
line-based protocol:

  <timestamp>;<cluster>;<operation>;<message>;

timestamp is the number of microseconds since epoch; cluster defines
either "new", "old" or "none"; operation defines what actually did
happen and message contains the user defined message per report. The
timestamp is monotonically increasing for each operation such that
reading files in order is less important, operations can still be
individually sorted.

This commit includes a small set of progress report operations, while
not likely to be the final set it's enough to show how more can be
added.

Discussion:
  https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/CVgK36oU5Oo