Fix gpcheckmirrorseg.pl query after introducing relfilenode counter
A query in gpcheckmirrorseg.pl would lookup relfilenodes obtained from
physically diffing segment files between primaries and mirrors. The
issue with the query was that it only looked at master's catalog which
would most likely not contain the relfilenodes of the QE
segments... especially after introducing the relfilenode counter
change. We union all the QE segment catalogs with master's catalog and
check by content id to correct the query.