Mask columns exceeding threshold in analyze sample
Analyze collects a sample from the table, in case the
sample contains columns with huge length, the columns may
result memory usage to go high cancelling the query.

This commit masks the lengthy columns to avoid high
memory usage while collecting sample. If a column has
exceeded a threshold no stats will be collected for this
column

Approach to address this issue :

 Numeric types or Fixed length user defined types:
  Always include the column in the sample.

 Built in Variable length type :
  Use octet_length() to get the detoasted size and exclude the column if the size exceeds column_threshhold (uncompressed)

 User defined variable length type :
  Check the toasted/compressed size and exclude the column if size exceeds column_threshhold (compressed)

Signed-off-by: Omer Arap <oarap@pivotal.io>