Fix bug in duplicate distribution columns for CTAS
When creating a table using the CTAS construction with an explicit
distribution clause, the duplicate check for distribution columns
wasn't applied:

  CREATE TABLE tt AS SELECT * FROM t DISTRIBUTED BY (c,c);

Move the duplicate distribution check to the parsing stage since
it is a syntax error to supply duplicate column references. This
alters the behavior such that supplying a duplicate non-existing
column will error out on the duplication rather than the column
being non-existing. Since syntax errors should have precedence
over semantic errors I however find that this is the correct order.

Remove the existing duplicate checks which only applied to CREATE
TABLE and ALTER TABLE, and add test cases to cover CTAS. The ticket
reference was removed since it's not public and also clearly didn't
cover all the cases. The DROP TABLE is also moved to not using IF
EXISTS since it would be an error if the table didn't exist at that
point.