Support COPY ON SEGMENT command
Support COPY statement that exports the table directly from segment
to local file parallelly.

This commit adds a keyword "on segment" to save the copied file on
"segment" instead of on "master".

Two place holders are used, which are "<SEG_DATA_DIR>" and "<SEGID>"
and will be replced to segment datadir and segment id.

E.g.

```
COPY tbl TO '/tmp/<SEG_DATA_DIR>filename<SEGID>.txt' ON SEGMENT;
```

Signed-off-by: Yuan Zhao <yuzhao@pivotal.io>
Signed-off-by: Haozhou Wang <hawang@pivotal.io>
Signed-off-by: Adam Lee <ali@pivotal.io>