Add test case for a query that used to not do DPE with old implementation.
The old code could not do partition elimination for this query, because
it looked for the pattern:

Join
  Append
  <other>

While in this query, the plan looks like:

Join
  Another Join
    Append
  <other>

The new code can apply the elimination also when the Append node is not
a direct child of the Join node, so we will now do partition elimination
for this query.

From the code point of view, this case isn't any different from when the
Append is a direct child of the join, so there's little reason to believe
this wouldn't work. But better safe than sorry, and add a test case for
it.