Remove underutilized CdbCellbuf facility.
It was only used in hash aggregates, and it only used a fraction of the
functionality. In essence, it was using it as a very simple memory
allocator, where each allocation was fixed size, and the only way to free
was to reset the whole cellbuf. But the same code was using a different,
but similar, mpool_* mechanism for allocating other things stored in
the hash buckets. We might as well use mpool_alloc for the HashAggEntry
struct as well, and get rid of all the cellbuf code.