Allow simplification of functions returning a composite types with ORCA.
This was disabled a long time ago, but even after digging into the git
history, I could not figure out why. Nothing seems to break with it, and
if there's a problem with handling composite types in ORCA, disabling
this transformation would just be plastering over it anyway, because there
are many other ways that you could end up with composite type constants
in a query.