Move workfile manager unittest code to src/test/regress
The workfile_manager unittest code was previously included in all
builds, this is however functionality which shouldn't be in prod
installations. This moves all unit test code to src/test/regress
and links it to the regress object such that it can be tested with
the normal pg_regress process.

Add a random test from the suite into ICW just to exercise it, none
of these test were ever executed in automated suites before so not
sure what we actually want to run.