Fix HashStreamOpaque from bad access / double free'ing HashBitMap.
`HashBitMap` is owned by `BitmapIndexScanState`(`node->bitmap`).
`BitmapIndexScanState` will free the `HashBitMap` when `ExecEnd` is
called. So it is not the responsibility of `HashStreamOpaque` to free it
which just has the reference to it.
Also, it should never try to access
the `HashBitMap` when `HashStreamOpaque` gets freed. `HashBitMap` might
have been freed even before the `HashStreamOpaque` gets freed.