Performance tests for loading into AO/CO tables
Removing old defunct performance directory and replace with a simple
way to test AO/CO table loading. This will help measure performance
changes when we start generating XLOGs for AO/CO tables and implement
WAL replication between primaries and mirrors. This current
performance implementation is a bit hacky currently but it gets the
job done and does not rely on TPC (which we can't have in our source
unfortunately).

Current implementation:
1. Create csv data file (specified by NUM_COPIES Makefile variable)
2. Host the csv data file in an external table to load data into a
   base table faster (and to work around a known memory leak bug when
   doing many inserts in a LOOP)
3. Create the AO and AOCO tables and start testing loading times
   sequentially and concurrently
4. Parse the output into a csv to allow separate comparison analysis