Get rid of the in-memory hash table.
Rely on tuple locks on the AO segment table instead. When an AO segment
is chosen as insertion target, its AO segment table row is locked for
the transaction (by heap_lock_tuple, i.e. by stamping its XMAX). Later,
when the insertion finishes, the tuple is updated with the new EOF (no
change there).

This replaces the in-memory hash table. Since we're relying on the
xmin/xmax on the AO segment table row anyway, might as well double-down
on that mechanism.

This also gets rid of the special meaning of segment 0. Segment 0 is
now chosen as insertion target, like any other segment.

Signed-off-by: Heikki Linnakangas <hlinnakangas@pivotal.io>