It's not guaranteed that segfile #0 exists.
Vacuum might've removed it. This was actually a pre-existing bug, even
before all this refactoring. If you connected to a segment in utility mode,
and inserted to an AO table, the data went to segfile #0. If you then
deleted the data, and vacuumed the table, segfile #0 would be removed.
If you now connect in utility mode again, and try to insert data again,
it will fail because segfile #0 is missing.

I'm not sure it's a good idea to ever drop segfile #0. Doesn't that
screw up the backend code to choose a relfilenode for a new relation,
that checks that a relfilenode is not in use, by checking if the file
exists? I tried to truncate segfile #0 instead of dropping it, but I got
an error that the EOF in gp_persistent_relation_node was larger than
the file size. So some work would be needed to do that.

Signed-off-by: Heikki Linnakangas <hlinnakangas@pivotal.io>