gpconfig: Add flag --file-compare
The file postgresql.conf contains GUC settings which are applied upon GPDB restart.
--file-compare contrasts the current GPDB settings (stored in the database)
to settings in the file which would be applied upon restart.

For almost all GUCs, all segments should have the same value; the --file-compare flag
either summarizes success (when synchronized) or, in the case of
unexpected differences, outputs detailed information about a given GUC
on all segments.

Signed-off-by: Marbin Tan <mtan@pivotal.io>