Put a bandaid around xpath tests
On a vanilla CentOS, the xpath test would always fail due to some
patches Red Hat applies to libxml2.
The xpath test in general is fairly questionable: it largely overlaps in
coverage with the xml test from upstream, and the two offending tests
were only failing because of an error message that doesn't even come
from our (Postgres) code: it's an implementation detail of libxml2.
I am not fully comfortable ripping xpath out yet, but this patch should
make sure it passes on CentOS without vendoring a special version of
libxml2.