Revert "Avoid null pointer deref in partition selection"
Commit 7b6dc9bfff555e902a6a5b9333023a4c618de3fe was well-intentioned,
but wrong: in cases where `outerPlan` is NULL, inputSlot should be
`NULL`, but the execution of the partition selector should *NOT*
prematurely end there, instead, a static partition selection should
happen (once) before this operator was exhausted (returning NULL).

This is a pure revert commit. The `NULL` initialization will follow.

This reverts commit 7b6dc9bfff555e902a6a5b9333023a4c618de3fe.