Fix partition_locking test that fails intermittently
Objective of the test is only to check if DDL/DML on partition tables lock only
the root table on QD.  Child partitions are locked on QE as needed.  The
intermittent failures are due to more locks acquired than expected.  In all
such cases, the additional locks were AccessShare locks on some catalog table.
This commit filters out locks on catalog tables from expected output.

The intermittent failures merit investigation but that can happen regardless of
this test fix.