Always annotate outer/inner edges in the plan.
explain.pl annotated the edges for nestloop only. This commit expands
that to include any operator with multiple children. This way in the
case the left/right children are swapped, it is clearly visible.

The script also considers InitPlans as regular children, so they must
handled specially for labeling and be ignored. Also fixed the incorrect
assumption that all InitPlans belong to a different slice, but that is
not always the case.