Handle Greenplum 5.0 in heap page conversion
The heap page conversion is only applicable in upgrades from 4.3
to 5.0. Ensure that we aren't already on 5.0 when figuring out if
to convert. Also initialize the flag to false for extra safety.
Unless the queries find that we need to convert the underlying heap
pages we really shouldn't attempt it.