Add test script for running pg_upgrade in ICW
This adds a testrunner to pg_upgrade intended to run at the end of
ICW. The running gpdemo cluster is converted from the current GPDB
version to the same version, which shall result in an identical
cluster. The script first dumps the ICW cluster, the upgrades into
a new gpdemo cluster and diffs the dump from that with the original
dump. In case the cluster needs to be tweaked before the test, a
_pre.sql file can be supplied which will be executed against the
old cluster before dumping the schema of it. This file currently
drops the relations which hold constraints not yet supported by
pg_upgrade.

This is inspired by the upstream test runner for pg_upgrade.