Make orca_debug module compile
The orca_debug contrib module was in disrepair and no longer
compiled due to changes in the version strings. Fix the code
and also add the proper libraries to link against to give the
.so a fighting chance to compile. Override the compiler in the
Makefile to use a C++ compiler instead.

Functions already compiled in when building GPDB with Orca were
removed, missing prototypes added and an unused function removed.
This allows orca_debug to compile without warnings.