Remove dead code
There was a lot of unused code in the gpMgmt bash code that hasn't
been running for quite some time. On top of really dead code, some
codepaths were reachable, but useless (like printing a non-existing
version string). Remove the dead code and replace the version print
with working code which pulls the version from the canonical source.