Fix deadlock on master with ENTRY_DB_SINGLETON
Original deadlock is caused by a reader requesting a lock which is already held
by the writer of the same MPP session, and another session is waiting for a
conflicting mode on the same lock.

The fix is to avoid checking waitMask conflicts for reader (i.e. MyProc is
different from lockHolderProcPtr).

Detailed discussion of the deadlock issue is at:
https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/OS1-ODIK0P4/ZIzayBbMBwAJ

Signed-off-by: Xin Zhang <xzhang@pivotal.io>